Moved RecipeSearch.jsx into new client/containers folder and updated import statements throughout app to pull correct file location.
